Add test for too few arguments to pysu
